# More efficient projections for FAB