Hi!

As the title says, here are slightly more efficient projections used in FAB. The main differences for the implementation are:
- it does not use `nonzero()`;
- it does not move tensors from cpu to device (instead creates them directly on device);
- it avoids unnecessary cloning.

Cheers