@fra31 2020-11-24
Hi,

thanks for the contribution! I'll try to quickly have a look and merge it.
Do you roughly know which is the improvement in terms of time?