@jeromerony 2020-11-24
Sorry, I should have included them in the first place.

I ran a few benchmarks on a Titan RTX. First a synthetic benchmark for the projections alone (with cudnn.benchmark enabled). The results are the following (time in seconds) given the same data:
| Projection | Original | New   |
|------------|----------|-------|
| L1         | 1.79     | 0.522 |
| L2         | 0.565    | 0.35  |
| Linf       | 0.756    | 0.338 |

So the improvements depend on the projections. 

In term of run-time for FAB, I tried with the eval script (ResNet-18, 1000 samples of CIFAR10, Linf at 1/255) and, with the new projections, the attack is about 2% faster (317.6s instead of 323.6s) with cudnn.benchmark enabled. Obviously this difference is going to be smaller for larger models and vice versa.