@fra31 2020-11-24
Ok, thanks a lot, that's a significant improvement.

Yeah, for the full attack anyway the other computations should dominate in terms of time. But it makes definitely sense to have more efficient projections in general.