@jeromerony 2020-11-26
I made some other minor optimizations. It does not change much but it's a bit cleaner.