@fra31 2020-11-29
Ok, I ran some tests and the new projections seem equivalent to the old ones, but significantly faster.

Thanks again for the contribution!